third_party/blink/web_tests/platform/generic/fast/dom/wrapper-classes-expected.txt

This tests wrapper class names for JavaScript.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

Basics


PASS classString(document.createAttribute('x')) is "Attr"
FAIL classString(document.createAttribute('x').__proto__) should be AttrPrototype. Was Attr.
PASS classString(document.createAttribute('x').constructor) is "Function"
PASS document.createAttribute('x').constructor.name is "Attr"
PASS classString(document.createComment('')) is "Comment"
FAIL classString(document.createComment('').__proto__) should be CommentPrototype. Was Comment.
PASS classString(document.createComment('').constructor) is "Function"
PASS document.createComment('').constructor.name is "Comment"
PASS classString(document.createDocumentFragment()) is "DocumentFragment"
FAIL classString(document.createDocumentFragment().__proto__) should be DocumentFragmentPrototype. Was DocumentFragment.
PASS classString(document.createDocumentFragment().constructor) is "Function"
PASS document.createDocumentFragment().constructor.name is "DocumentFragment"
PASS classString(document.implementation) is "DOMImplementation"
FAIL classString(document.implementation.__proto__) should be DOMImplementationPrototype. Was DOMImplementation.
PASS classString(document.implementation.constructor) is "Function"
PASS document.implementation.constructor.name is "DOMImplementation"
PASS classString(root.attributes) is "NamedNodeMap"
FAIL classString(root.attributes.__proto__) should be NamedNodeMapPrototype. Was NamedNodeMap.
PASS classString(root.attributes.constructor) is "Function"
PASS root.attributes.constructor.name is "NamedNodeMap"
PASS classString(document.createNodeIterator(root, 0, null, false)) is "NodeIterator"
FAIL classString(document.createNodeIterator(root, 0, null, false).__proto__) should be NodeIteratorPrototype. Was NodeIterator.
PASS classString(document.createNodeIterator(root, 0, null, false).constructor) is "Function"
PASS document.createNodeIterator(root, 0, null, false).constructor.name is "NodeIterator"
PASS classString(document.getElementsByTagName('div')) is "HTMLCollection"
FAIL classString(document.getElementsByTagName('div').__proto__) should be HTMLCollectionPrototype. Was HTMLCollection.
PASS classString(document.getElementsByTagName('div').constructor) is "Function"
PASS document.getElementsByTagName('div').constructor.name is "HTMLCollection"
PASS classString(document.createRange()) is "Range"
FAIL classString(document.createRange().__proto__) should be RangePrototype. Was Range.
PASS classString(document.createRange().constructor) is "Function"
PASS document.createRange().constructor.name is "Range"
PASS classString(document.createTextNode('')) is "Text"
FAIL classString(document.createTextNode('').__proto__) should be TextPrototype. Was Text.
PASS classString(document.createTextNode('').constructor) is "Function"
PASS document.createTextNode('').constructor.name is "Text"
PASS classString(document.createTreeWalker(root, 0, null, false)) is "TreeWalker"
FAIL classString(document.createTreeWalker(root, 0, null, false).__proto__) should be TreeWalkerPrototype. Was TreeWalker.
PASS classString(document.createTreeWalker(root, 0, null, false).constructor) is "Function"
PASS document.createTreeWalker(root, 0, null, false).constructor.name is "TreeWalker"


XML


PASS classString(xmlDocument) is "XMLDocument"
FAIL classString(xmlDocument.__proto__) should be XMLDocumentPrototype. Was XMLDocument.
PASS classString(xmlDocument.constructor) is "Function"
PASS xmlDocument.constructor.name is "XMLDocument"
PASS classString(xmlDocument.createCDATASection('')) is "CDATASection"
FAIL classString(xmlDocument.createCDATASection('').__proto__) should be CDATASectionPrototype. Was CDATASection.
PASS classString(xmlDocument.createCDATASection('').constructor) is "Function"
PASS xmlDocument.createCDATASection('').constructor.name is "CDATASection"
PASS classString(xmlDocument.createElementNS('x', 'x', 'x')) is "Element"
FAIL classString(xmlDocument.createElementNS('x', 'x', 'x').__proto__) should be ElementPrototype. Was Element.
PASS classString(xmlDocument.createElementNS('x', 'x', 'x').constructor) is "Function"
PASS xmlDocument.createElementNS('x', 'x', 'x').constructor.name is "Element"
PASS classString(xmlDocument.createProcessingInstruction('x', '')) is "ProcessingInstruction"
FAIL classString(xmlDocument.createProcessingInstruction('x', '').__proto__) should be ProcessingInstructionPrototype. Was ProcessingInstruction.
PASS classString(xmlDocument.createProcessingInstruction('x', '').constructor) is "Function"
PASS xmlDocument.createProcessingInstruction('x', '').constructor.name is "ProcessingInstruction"


Events


PASS classString(document.createEvent('Event')) is "Event"
FAIL classString(document.createEvent('Event').__proto__) should be EventPrototype. Was Event.
PASS classString(document.createEvent('Event').constructor) is "Function"
PASS document.createEvent('Event').constructor.name is "Event"
PASS classString(document.createEvent('KeyboardEvent')) is "KeyboardEvent"
FAIL classString(document.createEvent('KeyboardEvent').__proto__) should be KeyboardEventPrototype. Was KeyboardEvent.
PASS classString(document.createEvent('KeyboardEvent').constructor) is "Function"
PASS document.createEvent('KeyboardEvent').constructor.name is "KeyboardEvent"
PASS classString(document.createEvent('MouseEvent')) is "MouseEvent"
FAIL classString(document.createEvent('MouseEvent').__proto__) should be MouseEventPrototype. Was MouseEvent.
PASS classString(document.createEvent('MouseEvent').constructor) is "Function"
PASS document.createEvent('MouseEvent').constructor.name is "MouseEvent"
PASS classString(document.createEvent('MutationEvent')) is "MutationEvent"
FAIL classString(document.createEvent('MutationEvent').__proto__) should be MutationEventPrototype. Was MutationEvent.
PASS classString(document.createEvent('MutationEvent').constructor) is "Function"
PASS document.createEvent('MutationEvent').constructor.name is "MutationEvent"
PASS classString(document.createEvent('TextEvent')) is "TextEvent"
FAIL classString(document.createEvent('TextEvent').__proto__) should be TextEventPrototype. Was TextEvent.
PASS classString(document.createEvent('TextEvent').constructor) is "Function"
PASS document.createEvent('TextEvent').constructor.name is "TextEvent"
PASS classString(document.createEvent('UIEvent')) is "UIEvent"
FAIL classString(document.createEvent('UIEvent').__proto__) should be UIEventPrototype. Was UIEvent.
PASS classString(document.createEvent('UIEvent').constructor) is "Function"
PASS document.createEvent('UIEvent').constructor.name is "UIEvent"
PASS classString(document.createEvent('WheelEvent')) is "WheelEvent"
FAIL classString(document.createEvent('WheelEvent').__proto__) should be WheelEventPrototype. Was WheelEvent.
PASS classString(document.createEvent('WheelEvent').constructor) is "Function"
PASS document.createEvent('WheelEvent').constructor.name is "WheelEvent"


CSS DOM


PASS classString(document.styleSheets) is "StyleSheetList"
FAIL classString(document.styleSheets.__proto__) should be StyleSheetListPrototype. Was StyleSheetList.
PASS classString(document.styleSheets.constructor) is "Function"
PASS document.styleSheets.constructor.name is "StyleSheetList"
PASS classString(stylesheet) is "CSSStyleSheet"
FAIL classString(stylesheet.__proto__) should be CSSStyleSheetPrototype. Was CSSStyleSheet.
PASS classString(stylesheet.constructor) is "Function"
PASS stylesheet.constructor.name is "CSSStyleSheet"
PASS classString(stylesheet.cssRules) is "CSSRuleList"
FAIL classString(stylesheet.cssRules.__proto__) should be CSSRuleListPrototype. Was CSSRuleList.
PASS classString(stylesheet.cssRules.constructor) is "Function"
PASS stylesheet.cssRules.constructor.name is "CSSRuleList"
PASS classString(stylesheet.cssRules.item(0)) is "CSSStyleRule"
FAIL classString(stylesheet.cssRules.item(0).__proto__) should be CSSStyleRulePrototype. Was CSSStyleRule.
PASS classString(stylesheet.cssRules.item(0).constructor) is "Function"
PASS stylesheet.cssRules.item(0).constructor.name is "CSSStyleRule"
PASS classString(stylesheet.cssRules.item(0).style) is "CSSStyleDeclaration"
FAIL classString(stylesheet.cssRules.item(0).style.__proto__) should be CSSStyleDeclarationPrototype. Was CSSStyleDeclaration.
PASS classString(stylesheet.cssRules.item(0).style.constructor) is "Function"
PASS stylesheet.cssRules.item(0).style.constructor.name is "CSSStyleDeclaration"
PASS classString(stylesheet.media) is "MediaList"
FAIL classString(stylesheet.media.__proto__) should be MediaListPrototype. Was MediaList.
PASS classString(stylesheet.media.constructor) is "Function"
PASS stylesheet.media.constructor.name is "MediaList"


XPath


PASS classString(document.createExpression('//*', document.createNSResolver(document))) is "XPathExpression"
FAIL classString(document.createExpression('//*', document.createNSResolver(document)).__proto__) should be XPathExpressionPrototype. Was XPathExpression.
PASS classString(document.createExpression('//*', document.createNSResolver(document)).constructor) is "Function"
PASS document.createExpression('//*', document.createNSResolver(document)).constructor.name is "XPathExpression"
PASS classString(document.evaluate('//*', document, document.createNSResolver(document), 0, null)) is "XPathResult"
FAIL classString(document.evaluate('//*', document, document.createNSResolver(document), 0, null).__proto__) should be XPathResultPrototype. Was XPathResult.
PASS classString(document.evaluate('//*', document, document.createNSResolver(document), 0, null).constructor) is "Function"
PASS document.evaluate('//*', document, document.createNSResolver(document), 0, null).constructor.name is "XPathResult"


Other


PASS classString(window) is "Window"
FAIL classString(window.__proto__) should be WindowPrototype. Was Window.
PASS classString(window.constructor) is "Function"
PASS window.constructor.name is "Window"


HTML DOM


PASS classString(root.children) is "HTMLCollection"
FAIL classString(root.children.__proto__) should be HTMLCollectionPrototype. Was HTMLCollection.
PASS classString(root.children.constructor) is "Function"
PASS root.children.constructor.name is "HTMLCollection"
PASS classString(document) is "HTMLDocument"
FAIL classString(document.__proto__) should be HTMLDocumentPrototype. Was HTMLDocument.
PASS classString(document.constructor) is "Function"
PASS document.constructor.name is "HTMLDocument"
PASS classString(document.createElement('select').options) is "HTMLOptionsCollection"
FAIL classString(document.createElement('select').options.__proto__) should be HTMLOptionsCollectionPrototype. Was HTMLOptionsCollection.
PASS classString(document.createElement('select').options.constructor) is "Function"
PASS document.createElement('select').options.constructor.name is "HTMLOptionsCollection"


HTML Elements


PASS tagClassString('a') is "HTMLAnchorElement"
FAIL tagPrototypeClassString('a') should be HTMLAnchorElementPrototype. Was HTMLAnchorElement.
PASS tagConstructorClassString('a') is "Function"
PASS tagConstructorName('a') is "HTMLAnchorElement"
PASS tagClassString('abbr') is "HTMLElement"
FAIL tagPrototypeClassString('abbr') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('abbr') is "Function"
PASS tagConstructorName('abbr') is "HTMLElement"
PASS tagClassString('acronym') is "HTMLElement"
FAIL tagPrototypeClassString('acronym') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('acronym') is "Function"
PASS tagConstructorName('acronym') is "HTMLElement"
PASS tagClassString('address') is "HTMLElement"
FAIL tagPrototypeClassString('address') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('address') is "Function"
PASS tagConstructorName('address') is "HTMLElement"
PASS tagClassString('area') is "HTMLAreaElement"
FAIL tagPrototypeClassString('area') should be HTMLAreaElementPrototype. Was HTMLAreaElement.
PASS tagConstructorClassString('area') is "Function"
PASS tagConstructorName('area') is "HTMLAreaElement"
PASS tagClassString('article') is "HTMLElement"
FAIL tagPrototypeClassString('article') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('article') is "Function"
PASS tagConstructorName('article') is "HTMLElement"
PASS tagClassString('aside') is "HTMLElement"
FAIL tagPrototypeClassString('aside') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('aside') is "Function"
PASS tagConstructorName('aside') is "HTMLElement"
PASS tagClassString('b') is "HTMLElement"
FAIL tagPrototypeClassString('b') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('b') is "Function"
PASS tagConstructorName('b') is "HTMLElement"
PASS tagClassString('base') is "HTMLBaseElement"
FAIL tagPrototypeClassString('base') should be HTMLBaseElementPrototype. Was HTMLBaseElement.
PASS tagConstructorClassString('base') is "Function"
PASS tagConstructorName('base') is "HTMLBaseElement"
PASS tagClassString('basefont') is "HTMLElement"
FAIL tagPrototypeClassString('basefont') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('basefont') is "Function"
PASS tagConstructorName('basefont') is "HTMLElement"
PASS tagClassString('bdo') is "HTMLElement"
FAIL tagPrototypeClassString('bdo') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('bdo') is "Function"
PASS tagConstructorName('bdo') is "HTMLElement"
PASS tagClassString('bgsound') is "HTMLUnknownElement"
FAIL tagPrototypeClassString('bgsound') should be HTMLUnknownElementPrototype. Was HTMLUnknownElement.
PASS tagConstructorClassString('bgsound') is "Function"
PASS tagConstructorName('bgsound') is "HTMLUnknownElement"
PASS tagClassString('big') is "HTMLElement"
FAIL tagPrototypeClassString('big') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('big') is "Function"
PASS tagConstructorName('big') is "HTMLElement"
PASS tagClassString('blockquote') is "HTMLQuoteElement"
FAIL tagPrototypeClassString('blockquote') should be HTMLQuoteElementPrototype. Was HTMLQuoteElement.
PASS tagConstructorClassString('blockquote') is "Function"
PASS tagConstructorName('blockquote') is "HTMLQuoteElement"
PASS tagClassString('body') is "HTMLBodyElement"
FAIL tagPrototypeClassString('body') should be HTMLBodyElementPrototype. Was HTMLBodyElement.
PASS tagConstructorClassString('body') is "Function"
PASS tagConstructorName('body') is "HTMLBodyElement"
PASS tagClassString('br') is "HTMLBRElement"
FAIL tagPrototypeClassString('br') should be HTMLBRElementPrototype. Was HTMLBRElement.
PASS tagConstructorClassString('br') is "Function"
PASS tagConstructorName('br') is "HTMLBRElement"
PASS tagClassString('button') is "HTMLButtonElement"
FAIL tagPrototypeClassString('button') should be HTMLButtonElementPrototype. Was HTMLButtonElement.
PASS tagConstructorClassString('button') is "Function"
PASS tagConstructorName('button') is "HTMLButtonElement"
PASS tagClassString('canvas') is "HTMLCanvasElement"
FAIL tagPrototypeClassString('canvas') should be HTMLCanvasElementPrototype. Was HTMLCanvasElement.
PASS tagConstructorClassString('canvas') is "Function"
PASS tagConstructorName('canvas') is "HTMLCanvasElement"
PASS tagClassString('caption') is "HTMLTableCaptionElement"
FAIL tagPrototypeClassString('caption') should be HTMLTableCaptionElementPrototype. Was HTMLTableCaptionElement.
PASS tagConstructorClassString('caption') is "Function"
PASS tagConstructorName('caption') is "HTMLTableCaptionElement"
PASS tagClassString('center') is "HTMLElement"
FAIL tagPrototypeClassString('center') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('center') is "Function"
PASS tagConstructorName('center') is "HTMLElement"
PASS tagClassString('cite') is "HTMLElement"
FAIL tagPrototypeClassString('cite') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('cite') is "Function"
PASS tagConstructorName('cite') is "HTMLElement"
PASS tagClassString('code') is "HTMLElement"
FAIL tagPrototypeClassString('code') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('code') is "Function"
PASS tagConstructorName('code') is "HTMLElement"
PASS tagClassString('col') is "HTMLTableColElement"
FAIL tagPrototypeClassString('col') should be HTMLTableColElementPrototype. Was HTMLTableColElement.
PASS tagConstructorClassString('col') is "Function"
PASS tagConstructorName('col') is "HTMLTableColElement"
PASS tagClassString('colgroup') is "HTMLTableColElement"
FAIL tagPrototypeClassString('colgroup') should be HTMLTableColElementPrototype. Was HTMLTableColElement.
PASS tagConstructorClassString('colgroup') is "Function"
PASS tagConstructorName('colgroup') is "HTMLTableColElement"
PASS tagClassString('dd') is "HTMLElement"
FAIL tagPrototypeClassString('dd') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('dd') is "Function"
PASS tagConstructorName('dd') is "HTMLElement"
PASS tagClassString('del') is "HTMLModElement"
FAIL tagPrototypeClassString('del') should be HTMLModElementPrototype. Was HTMLModElement.
PASS tagConstructorClassString('del') is "Function"
PASS tagConstructorName('del') is "HTMLModElement"
PASS tagClassString('dfn') is "HTMLElement"
FAIL tagPrototypeClassString('dfn') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('dfn') is "Function"
PASS tagConstructorName('dfn') is "HTMLElement"
PASS tagClassString('dir') is "HTMLDirectoryElement"
FAIL tagPrototypeClassString('dir') should be HTMLDirectoryElementPrototype. Was HTMLDirectoryElement.
PASS tagConstructorClassString('dir') is "Function"
PASS tagConstructorName('dir') is "HTMLDirectoryElement"
PASS tagClassString('div') is "HTMLDivElement"
FAIL tagPrototypeClassString('div') should be HTMLDivElementPrototype. Was HTMLDivElement.
PASS tagConstructorClassString('div') is "Function"
PASS tagConstructorName('div') is "HTMLDivElement"
PASS tagClassString('dl') is "HTMLDListElement"
FAIL tagPrototypeClassString('dl') should be HTMLDListElementPrototype. Was HTMLDListElement.
PASS tagConstructorClassString('dl') is "Function"
PASS tagConstructorName('dl') is "HTMLDListElement"
PASS tagClassString('dt') is "HTMLElement"
FAIL tagPrototypeClassString('dt') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('dt') is "Function"
PASS tagConstructorName('dt') is "HTMLElement"
PASS tagClassString('em') is "HTMLElement"
FAIL tagPrototypeClassString('em') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('em') is "Function"
PASS tagConstructorName('em') is "HTMLElement"
PASS tagClassString('embed') is "HTMLEmbedElement"
FAIL tagPrototypeClassString('embed') should be HTMLEmbedElementPrototype. Was HTMLEmbedElement.
PASS tagConstructorClassString('embed') is "Function"
PASS tagConstructorName('embed') is "HTMLEmbedElement"
PASS tagClassString('fieldset') is "HTMLFieldSetElement"
FAIL tagPrototypeClassString('fieldset') should be HTMLFieldSetElementPrototype. Was HTMLFieldSetElement.
PASS tagConstructorClassString('fieldset') is "Function"
PASS tagConstructorName('fieldset') is "HTMLFieldSetElement"
PASS tagClassString('font') is "HTMLFontElement"
FAIL tagPrototypeClassString('font') should be HTMLFontElementPrototype. Was HTMLFontElement.
PASS tagConstructorClassString('font') is "Function"
PASS tagConstructorName('font') is "HTMLFontElement"
PASS tagClassString('footer') is "HTMLElement"
FAIL tagPrototypeClassString('footer') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('footer') is "Function"
PASS tagConstructorName('footer') is "HTMLElement"
PASS tagClassString('form') is "HTMLFormElement"
FAIL tagPrototypeClassString('form') should be HTMLFormElementPrototype. Was HTMLFormElement.
PASS tagConstructorClassString('form') is "Function"
PASS tagConstructorName('form') is "HTMLFormElement"
PASS tagClassString('frame') is "HTMLFrameElement"
FAIL tagPrototypeClassString('frame') should be HTMLFrameElementPrototype. Was HTMLFrameElement.
PASS tagConstructorClassString('frame') is "Function"
PASS tagConstructorName('frame') is "HTMLFrameElement"
PASS tagClassString('frameset') is "HTMLFrameSetElement"
FAIL tagPrototypeClassString('frameset') should be HTMLFrameSetElementPrototype. Was HTMLFrameSetElement.
PASS tagConstructorClassString('frameset') is "Function"
PASS tagConstructorName('frameset') is "HTMLFrameSetElement"
PASS tagClassString('head') is "HTMLHeadElement"
FAIL tagPrototypeClassString('head') should be HTMLHeadElementPrototype. Was HTMLHeadElement.
PASS tagConstructorClassString('head') is "Function"
PASS tagConstructorName('head') is "HTMLHeadElement"
PASS tagClassString('h1') is "HTMLHeadingElement"
FAIL tagPrototypeClassString('h1') should be HTMLHeadingElementPrototype. Was HTMLHeadingElement.
PASS tagConstructorClassString('h1') is "Function"
PASS tagConstructorName('h1') is "HTMLHeadingElement"
PASS tagClassString('h2') is "HTMLHeadingElement"
FAIL tagPrototypeClassString('h2') should be HTMLHeadingElementPrototype. Was HTMLHeadingElement.
PASS tagConstructorClassString('h2') is "Function"
PASS tagConstructorName('h2') is "HTMLHeadingElement"
PASS tagClassString('h3') is "HTMLHeadingElement"
FAIL tagPrototypeClassString('h3') should be HTMLHeadingElementPrototype. Was HTMLHeadingElement.
PASS tagConstructorClassString('h3') is "Function"
PASS tagConstructorName('h3') is "HTMLHeadingElement"
PASS tagClassString('h4') is "HTMLHeadingElement"
FAIL tagPrototypeClassString('h4') should be HTMLHeadingElementPrototype. Was HTMLHeadingElement.
PASS tagConstructorClassString('h4') is "Function"
PASS tagConstructorName('h4') is "HTMLHeadingElement"
PASS tagClassString('h5') is "HTMLHeadingElement"
FAIL tagPrototypeClassString('h5') should be HTMLHeadingElementPrototype. Was HTMLHeadingElement.
PASS tagConstructorClassString('h5') is "Function"
PASS tagConstructorName('h5') is "HTMLHeadingElement"
PASS tagClassString('h6') is "HTMLHeadingElement"
FAIL tagPrototypeClassString('h6') should be HTMLHeadingElementPrototype. Was HTMLHeadingElement.
PASS tagConstructorClassString('h6') is "Function"
PASS tagConstructorName('h6') is "HTMLHeadingElement"
PASS tagClassString('head') is "HTMLHeadElement"
FAIL tagPrototypeClassString('head') should be HTMLHeadElementPrototype. Was HTMLHeadElement.
PASS tagConstructorClassString('head') is "Function"
PASS tagConstructorName('head') is "HTMLHeadElement"
PASS tagClassString('header') is "HTMLElement"
FAIL tagPrototypeClassString('header') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('header') is "Function"
PASS tagConstructorName('header') is "HTMLElement"
PASS tagClassString('hgroup') is "HTMLElement"
FAIL tagPrototypeClassString('hgroup') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('hgroup') is "Function"
PASS tagConstructorName('hgroup') is "HTMLElement"
PASS tagClassString('hr') is "HTMLHRElement"
FAIL tagPrototypeClassString('hr') should be HTMLHRElementPrototype. Was HTMLHRElement.
PASS tagConstructorClassString('hr') is "Function"
PASS tagConstructorName('hr') is "HTMLHRElement"
PASS tagClassString('html') is "HTMLHtmlElement"
FAIL tagPrototypeClassString('html') should be HTMLHtmlElementPrototype. Was HTMLHtmlElement.
PASS tagConstructorClassString('html') is "Function"
PASS tagConstructorName('html') is "HTMLHtmlElement"
PASS tagClassString('i') is "HTMLElement"
FAIL tagPrototypeClassString('i') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('i') is "Function"
PASS tagConstructorName('i') is "HTMLElement"
PASS tagClassString('iframe') is "HTMLIFrameElement"
FAIL tagPrototypeClassString('iframe') should be HTMLIFrameElementPrototype. Was HTMLIFrameElement.
PASS tagConstructorClassString('iframe') is "Function"
PASS tagConstructorName('iframe') is "HTMLIFrameElement"
PASS tagClassString('image') is "HTMLUnknownElement"
FAIL tagPrototypeClassString('image') should be HTMLUnknownElementPrototype. Was HTMLUnknownElement.
PASS tagConstructorClassString('image') is "Function"
PASS tagConstructorName('image') is "HTMLUnknownElement"
PASS tagClassString('img') is "HTMLImageElement"
FAIL tagPrototypeClassString('img') should be HTMLImageElementPrototype. Was HTMLImageElement.
PASS tagConstructorClassString('img') is "Function"
PASS tagConstructorName('img') is "HTMLImageElement"
PASS tagClassString('input') is "HTMLInputElement"
FAIL tagPrototypeClassString('input') should be HTMLInputElementPrototype. Was HTMLInputElement.
PASS tagConstructorClassString('input') is "Function"
PASS tagConstructorName('input') is "HTMLInputElement"
PASS tagClassString('ins') is "HTMLModElement"
FAIL tagPrototypeClassString('ins') should be HTMLModElementPrototype. Was HTMLModElement.
PASS tagConstructorClassString('ins') is "Function"
PASS tagConstructorName('ins') is "HTMLModElement"
PASS tagClassString('kbd') is "HTMLElement"
FAIL tagPrototypeClassString('kbd') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('kbd') is "Function"
PASS tagConstructorName('kbd') is "HTMLElement"
PASS tagClassString('label') is "HTMLLabelElement"
FAIL tagPrototypeClassString('label') should be HTMLLabelElementPrototype. Was HTMLLabelElement.
PASS tagConstructorClassString('label') is "Function"
PASS tagConstructorName('label') is "HTMLLabelElement"
PASS tagClassString('layer') is "HTMLElement"
FAIL tagPrototypeClassString('layer') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('layer') is "Function"
PASS tagConstructorName('layer') is "HTMLElement"
PASS tagClassString('legend') is "HTMLLegendElement"
FAIL tagPrototypeClassString('legend') should be HTMLLegendElementPrototype. Was HTMLLegendElement.
PASS tagConstructorClassString('legend') is "Function"
PASS tagConstructorName('legend') is "HTMLLegendElement"
PASS tagClassString('li') is "HTMLLIElement"
FAIL tagPrototypeClassString('li') should be HTMLLIElementPrototype. Was HTMLLIElement.
PASS tagConstructorClassString('li') is "Function"
PASS tagConstructorName('li') is "HTMLLIElement"
PASS tagClassString('link') is "HTMLLinkElement"
FAIL tagPrototypeClassString('link') should be HTMLLinkElementPrototype. Was HTMLLinkElement.
PASS tagConstructorClassString('link') is "Function"
PASS tagConstructorName('link') is "HTMLLinkElement"
PASS tagClassString('listing') is "HTMLPreElement"
FAIL tagPrototypeClassString('listing') should be HTMLPreElementPrototype. Was HTMLPreElement.
PASS tagConstructorClassString('listing') is "Function"
PASS tagConstructorName('listing') is "HTMLPreElement"
PASS tagClassString('main') is "HTMLElement"
FAIL tagPrototypeClassString('main') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('main') is "Function"
PASS tagConstructorName('main') is "HTMLElement"
PASS tagClassString('map') is "HTMLMapElement"
FAIL tagPrototypeClassString('map') should be HTMLMapElementPrototype. Was HTMLMapElement.
PASS tagConstructorClassString('map') is "Function"
PASS tagConstructorName('map') is "HTMLMapElement"
PASS tagClassString('marquee') is "HTMLMarqueeElement"
FAIL tagPrototypeClassString('marquee') should be HTMLMarqueeElementPrototype. Was HTMLMarqueeElement.
PASS tagConstructorClassString('marquee') is "Function"
PASS tagConstructorName('marquee') is "HTMLMarqueeElement"
PASS tagClassString('menu') is "HTMLMenuElement"
FAIL tagPrototypeClassString('menu') should be HTMLMenuElementPrototype. Was HTMLMenuElement.
PASS tagConstructorClassString('menu') is "Function"
PASS tagConstructorName('menu') is "HTMLMenuElement"
PASS tagClassString('meta') is "HTMLMetaElement"
FAIL tagPrototypeClassString('meta') should be HTMLMetaElementPrototype. Was HTMLMetaElement.
PASS tagConstructorClassString('meta') is "Function"
PASS tagConstructorName('meta') is "HTMLMetaElement"
PASS tagClassString('nav') is "HTMLElement"
FAIL tagPrototypeClassString('nav') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('nav') is "Function"
PASS tagConstructorName('nav') is "HTMLElement"
PASS tagClassString('nobr') is "HTMLElement"
FAIL tagPrototypeClassString('nobr') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('nobr') is "Function"
PASS tagConstructorName('nobr') is "HTMLElement"
PASS tagClassString('noembed') is "HTMLElement"
FAIL tagPrototypeClassString('noembed') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('noembed') is "Function"
PASS tagConstructorName('noembed') is "HTMLElement"
PASS tagClassString('noframes') is "HTMLElement"
FAIL tagPrototypeClassString('noframes') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('noframes') is "Function"
PASS tagConstructorName('noframes') is "HTMLElement"
PASS tagClassString('nolayer') is "HTMLElement"
FAIL tagPrototypeClassString('nolayer') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('nolayer') is "Function"
PASS tagConstructorName('nolayer') is "HTMLElement"
PASS tagClassString('noscript') is "HTMLElement"
FAIL tagPrototypeClassString('noscript') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('noscript') is "Function"
PASS tagConstructorName('noscript') is "HTMLElement"
PASS tagClassString('object') is "HTMLObjectElement"
FAIL tagPrototypeClassString('object') should be HTMLObjectElementPrototype. Was HTMLObjectElement.
PASS tagConstructorClassString('object') is "Function"
PASS tagConstructorName('object') is "HTMLObjectElement"
PASS tagClassString('ol') is "HTMLOListElement"
FAIL tagPrototypeClassString('ol') should be HTMLOListElementPrototype. Was HTMLOListElement.
PASS tagConstructorClassString('ol') is "Function"
PASS tagConstructorName('ol') is "HTMLOListElement"
PASS tagClassString('optgroup') is "HTMLOptGroupElement"
FAIL tagPrototypeClassString('optgroup') should be HTMLOptGroupElementPrototype. Was HTMLOptGroupElement.
PASS tagConstructorClassString('optgroup') is "Function"
PASS tagConstructorName('optgroup') is "HTMLOptGroupElement"
PASS tagClassString('option') is "HTMLOptionElement"
FAIL tagPrototypeClassString('option') should be HTMLOptionElementPrototype. Was HTMLOptionElement.
PASS tagConstructorClassString('option') is "Function"
PASS tagConstructorName('option') is "HTMLOptionElement"
PASS tagClassString('p') is "HTMLParagraphElement"
FAIL tagPrototypeClassString('p') should be HTMLParagraphElementPrototype. Was HTMLParagraphElement.
PASS tagConstructorClassString('p') is "Function"
PASS tagConstructorName('p') is "HTMLParagraphElement"
PASS tagClassString('param') is "HTMLParamElement"
FAIL tagPrototypeClassString('param') should be HTMLParamElementPrototype. Was HTMLParamElement.
PASS tagConstructorClassString('param') is "Function"
PASS tagConstructorName('param') is "HTMLParamElement"
PASS tagClassString('plaintext') is "HTMLElement"
FAIL tagPrototypeClassString('plaintext') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('plaintext') is "Function"
PASS tagConstructorName('plaintext') is "HTMLElement"
PASS tagClassString('pre') is "HTMLPreElement"
FAIL tagPrototypeClassString('pre') should be HTMLPreElementPrototype. Was HTMLPreElement.
PASS tagConstructorClassString('pre') is "Function"
PASS tagConstructorName('pre') is "HTMLPreElement"
PASS tagClassString('q') is "HTMLQuoteElement"
FAIL tagPrototypeClassString('q') should be HTMLQuoteElementPrototype. Was HTMLQuoteElement.
PASS tagConstructorClassString('q') is "Function"
PASS tagConstructorName('q') is "HTMLQuoteElement"
PASS tagClassString('rp') is "HTMLElement"
FAIL tagPrototypeClassString('rp') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('rp') is "Function"
PASS tagConstructorName('rp') is "HTMLElement"
PASS tagClassString('rt') is "HTMLElement"
FAIL tagPrototypeClassString('rt') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('rt') is "Function"
PASS tagConstructorName('rt') is "HTMLElement"
PASS tagClassString('ruby') is "HTMLElement"
FAIL tagPrototypeClassString('ruby') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('ruby') is "Function"
PASS tagConstructorName('ruby') is "HTMLElement"
PASS tagClassString('s') is "HTMLElement"
FAIL tagPrototypeClassString('s') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('s') is "Function"
PASS tagConstructorName('s') is "HTMLElement"
PASS tagClassString('samp') is "HTMLElement"
FAIL tagPrototypeClassString('samp') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('samp') is "Function"
PASS tagConstructorName('samp') is "HTMLElement"
PASS tagClassString('script') is "HTMLScriptElement"
FAIL tagPrototypeClassString('script') should be HTMLScriptElementPrototype. Was HTMLScriptElement.
PASS tagConstructorClassString('script') is "Function"
PASS tagConstructorName('script') is "HTMLScriptElement"
PASS tagClassString('section') is "HTMLElement"
FAIL tagPrototypeClassString('section') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('section') is "Function"
PASS tagConstructorName('section') is "HTMLElement"
PASS tagClassString('select') is "HTMLSelectElement"
FAIL tagPrototypeClassString('select') should be HTMLSelectElementPrototype. Was HTMLSelectElement.
PASS tagConstructorClassString('select') is "Function"
PASS tagConstructorName('select') is "HTMLSelectElement"
PASS tagClassString('small') is "HTMLElement"
FAIL tagPrototypeClassString('small') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('small') is "Function"
PASS tagConstructorName('small') is "HTMLElement"
PASS tagClassString('span') is "HTMLSpanElement"
FAIL tagPrototypeClassString('span') should be HTMLSpanElementPrototype. Was HTMLSpanElement.
PASS tagConstructorClassString('span') is "Function"
PASS tagConstructorName('span') is "HTMLSpanElement"
PASS tagClassString('strike') is "HTMLElement"
FAIL tagPrototypeClassString('strike') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('strike') is "Function"
PASS tagConstructorName('strike') is "HTMLElement"
PASS tagClassString('strong') is "HTMLElement"
FAIL tagPrototypeClassString('strong') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('strong') is "Function"
PASS tagConstructorName('strong') is "HTMLElement"
PASS tagClassString('style') is "HTMLStyleElement"
FAIL tagPrototypeClassString('style') should be HTMLStyleElementPrototype. Was HTMLStyleElement.
PASS tagConstructorClassString('style') is "Function"
PASS tagConstructorName('style') is "HTMLStyleElement"
PASS tagClassString('sub') is "HTMLElement"
FAIL tagPrototypeClassString('sub') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('sub') is "Function"
PASS tagConstructorName('sub') is "HTMLElement"
PASS tagClassString('sup') is "HTMLElement"
FAIL tagPrototypeClassString('sup') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('sup') is "Function"
PASS tagConstructorName('sup') is "HTMLElement"
PASS tagClassString('table') is "HTMLTableElement"
FAIL tagPrototypeClassString('table') should be HTMLTableElementPrototype. Was HTMLTableElement.
PASS tagConstructorClassString('table') is "Function"
PASS tagConstructorName('table') is "HTMLTableElement"
PASS tagClassString('tbody') is "HTMLTableSectionElement"
FAIL tagPrototypeClassString('tbody') should be HTMLTableSectionElementPrototype. Was HTMLTableSectionElement.
PASS tagConstructorClassString('tbody') is "Function"
PASS tagConstructorName('tbody') is "HTMLTableSectionElement"
PASS tagClassString('td') is "HTMLTableCellElement"
FAIL tagPrototypeClassString('td') should be HTMLTableCellElementPrototype. Was HTMLTableCellElement.
PASS tagConstructorClassString('td') is "Function"
PASS tagConstructorName('td') is "HTMLTableCellElement"
PASS tagClassString('textarea') is "HTMLTextAreaElement"
FAIL tagPrototypeClassString('textarea') should be HTMLTextAreaElementPrototype. Was HTMLTextAreaElement.
PASS tagConstructorClassString('textarea') is "Function"
PASS tagConstructorName('textarea') is "HTMLTextAreaElement"
PASS tagClassString('tfoot') is "HTMLTableSectionElement"
FAIL tagPrototypeClassString('tfoot') should be HTMLTableSectionElementPrototype. Was HTMLTableSectionElement.
PASS tagConstructorClassString('tfoot') is "Function"
PASS tagConstructorName('tfoot') is "HTMLTableSectionElement"
PASS tagClassString('th') is "HTMLTableCellElement"
FAIL tagPrototypeClassString('th') should be HTMLTableCellElementPrototype. Was HTMLTableCellElement.
PASS tagConstructorClassString('th') is "Function"
PASS tagConstructorName('th') is "HTMLTableCellElement"
PASS tagClassString('thead') is "HTMLTableSectionElement"
FAIL tagPrototypeClassString('thead') should be HTMLTableSectionElementPrototype. Was HTMLTableSectionElement.
PASS tagConstructorClassString('thead') is "Function"
PASS tagConstructorName('thead') is "HTMLTableSectionElement"
PASS tagClassString('title') is "HTMLTitleElement"
FAIL tagPrototypeClassString('title') should be HTMLTitleElementPrototype. Was HTMLTitleElement.
PASS tagConstructorClassString('title') is "Function"
PASS tagConstructorName('title') is "HTMLTitleElement"
PASS tagClassString('tr') is "HTMLTableRowElement"
FAIL tagPrototypeClassString('tr') should be HTMLTableRowElementPrototype. Was HTMLTableRowElement.
PASS tagConstructorClassString('tr') is "Function"
PASS tagConstructorName('tr') is "HTMLTableRowElement"
PASS tagClassString('tt') is "HTMLElement"
FAIL tagPrototypeClassString('tt') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('tt') is "Function"
PASS tagConstructorName('tt') is "HTMLElement"
PASS tagClassString('u') is "HTMLElement"
FAIL tagPrototypeClassString('u') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('u') is "Function"
PASS tagConstructorName('u') is "HTMLElement"
PASS tagClassString('ul') is "HTMLUListElement"
FAIL tagPrototypeClassString('ul') should be HTMLUListElementPrototype. Was HTMLUListElement.
PASS tagConstructorClassString('ul') is "Function"
PASS tagConstructorName('ul') is "HTMLUListElement"
PASS tagClassString('var') is "HTMLElement"
FAIL tagPrototypeClassString('var') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('var') is "Function"
PASS tagConstructorName('var') is "HTMLElement"
PASS tagClassString('wbr') is "HTMLElement"
FAIL tagPrototypeClassString('wbr') should be HTMLElementPrototype. Was HTMLElement.
PASS tagConstructorClassString('wbr') is "Function"
PASS tagConstructorName('wbr') is "HTMLElement"
PASS tagClassString('xmp') is "HTMLPreElement"
FAIL tagPrototypeClassString('xmp') should be HTMLPreElementPrototype. Was HTMLPreElement.
PASS tagConstructorClassString('xmp') is "Function"
PASS tagConstructorName('xmp') is "HTMLPreElement"
